refactor(eventStatusLogic): extract event lifecycle rules into pure logic
